+++ src/server/index.ts
@@ -12,4 +12,13 @@ export {
   type StarterServerOptions,
   type ServerStats,
   type MessageMeta,
+  type NackEvent,
 } from "./server.js";
+
+export {
+  buildRawAck,
+  resolveNackCode,
+  MllpAckError,
+  type AckCode,
+  type NegativeAckCode,
+} from "./ack.js";